CRM1

See merge request haptics/rm0!75

+++ source/Encoder/src/main.cpp
@@ -34,29 +34,52 @@
 #include <Encoder/include/AhapEncoder.h>
 #include <Encoder/include/IvsEncoder.h>
 #include <Encoder/include/PcmEncoder.h>
-#include <Types/include/IOJson.h>
+#include <IOHaptics/include/IOBinary.h>
+#include <IOHaptics/include/IOJson.h>
 #include <Tools/include/InputParser.h>
 #include <Tools/include/OHMData.h>
 #include <Types/include/Haptics.h>
 #include <Types/include/Perception.h>
-#include <functional>
 #include <filesystem>
+#include <functional>
 
 using haptics::encoder::AhapEncoder;
 using haptics::encoder::IvsEncoder;
 using haptics::encoder::PcmEncoder;
-using haptics::encoder::IOJson;
+using haptics::io::IOBinary;
+using haptics::io::IOJson;
 using haptics::tools::InputParser;
 using haptics::tools::OHMData;
 using haptics::types::Haptics;
 using haptics::types::Perception;
+
+auto help() -> void {
+  std::cout
+      << "usages: Encoder [-h] -f <FILE> -o <OUTPUT_FILE> [-b] [-kin]" << std::endl
+      << std::endl
+      << "This piece of software encodes an input file into a MPEG Haptics RM1 file" << std::endl
+      << "positional arguments:" << std::endl
+      << "\t-f, --file <FILE>\t\tfile to convert" << std::endl
+      << "\t-o, --output <OUTPUT_FILE>\toutput file" << std::endl
+      << std::endl
+      << "optional arguments:" << std::endl
+      << "\t-h, --help\t\t\tshow this help message and exit" << std::endl
+      << "\t-b, --binary\t\t\tthe file will be encoded into its binary format. If not provided "
+         "the encoder will output a file in a human-readable format"
+      << std::endl
+      << "\t-kb, --bitrate\t\t\ttarget bitrate of the encoded file" << std::endl
+      << "\t-kin, --kinesthetic\t\tIf provided, the file will be encoded as a kinesthetic effect. "
+         "Otherwise it will be considered as a vibrotactile effect (this option is currently "
+         "available only for wav files encoding)"
+      << std::endl;
+}
 
 // NOLINTNEXTLINE(bugprone-exception-escape, readability-function-size)
 auto main(int argc, char *argv[]) -> int {
   const auto args = std::vector<const char *>(argv, argv + argc);
   InputParser inputParser(args);
   if (inputParser.cmdOptionExists("-h") || inputParser.cmdOptionExists("--help")) {
-    InputParser::help(args[0]);
+    help();
     return EXIT_SUCCESS;
   }
 
@@ -65,7 +88,7 @@ auto main(int argc, char *argv[]) -> int {
     filename = inputParser.getCmdOption("--file");
   }
   if (filename.empty() || !std::filesystem::is_regular_file(filename)) {
-    InputParser::help(args[0]);
+    help();
     return EXIT_FAILURE;
   }
 
@@ -86,7 +109,8 @@ auto main(int argc, char *argv[]) -> int {
     std::cout << "The OHM file to process : " << filename << std::endl;
     OHMData ohmData;
     if (!ohmData.loadFile(filename)) {
-      std::cerr << "ERROR : impossible to read the OHM file : " << std::endl << filename << std::endl;
+      std::cerr << "ERROR : impossible to read the OHM file : " << std::endl
+                << filename << std::endl;
       return EXIT_FAILURE;
     }
     hapticFile.loadMetadataFromOHM(ohmData);
@@ -97,8 +121,9 @@ auto main(int argc, char *argv[]) -> int {
     }
     std::filesystem::path folderPath = std::filesystem::path(filename);
     folderPath = folderPath.parent_path();
-    for (int i = 0; i < ohmData.getHapticElementMetadataSize() && codeExit == EXIT_SUCCESS; i++) {
-      OHMData::HapticElementMetadata metadata = ohmData.getHapticElementMetadataAt(i);
+    for (uint32_t i = 0; i < ohmData.getHapticElementMetadataSize() && codeExit == EXIT_SUCCESS;
+         i++) {
+      OHMData::HapticElementMetadata metadata = ohmData.getHapticElementMetadataAt((int)i);
 
       filename = (folderPath / metadata.elementFilename).string();
       if (!std::filesystem::is_regular_file(filename)) {
@@ -107,7 +132,7 @@ auto main(int argc, char *argv[]) -> int {
       }
 
       ext = InputParser::getFileExt(filename);
-      myPerception = hapticFile.getPerceptionAt(i);
+      myPerception = hapticFile.getPerceptionAt((int)i);
       if (ext == "json" || ext == "ahap") {
         std::cout << "The AHAP file to encode : " << filename << std::endl;
         codeExit = AhapEncoder::encode(filename, myPerception);
@@ -117,17 +142,18 @@ auto main(int argc, char *argv[]) -> int {
       } else if (ext == "wav") {
         std::cout << "The WAV file to encode : " << filename << std::endl;
         haptics::encoder::EncodingConfig config;
-        // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-        config.curveFrequencyLimit = 72.5;
-        // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-        config.frequencyBandLimits = std::vector<std::pair<double, double>>{{72.5, 1000}};
-        // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-        config.windowLength = 15;
-        return PcmEncoder::encode(filename, config, myPerception);
+        if (inputParser.cmdOptionExists("-kb")) {
+          int bitrate = std::stoi(inputParser.getCmdOption("-kb"));
+          std::cout << "target bitrate: " << bitrate << " kb/s" << std::endl;
+          config = haptics::encoder::EncodingConfig::generateConfig(bitrate);
+        } else {
+          config = haptics::encoder::EncodingConfig::generateConfig();
+        }
+        codeExit = PcmEncoder::encode(filename, config, myPerception);
       }
 
       if (codeExit == EXIT_SUCCESS) {
-        hapticFile.replacePerceptionAt(i, myPerception);
+        hapticFile.replacePerceptionAt((int)i, myPerception);
       }
     }
   } else if (ext == "json" || ext == "ahap") {
@@ -141,28 +167,29 @@ auto main(int argc, char *argv[]) -> int {
   } else if (ext == "wav") {
     std::cout << "The WAV file to encode : " << filename << std::endl;
     haptics::encoder::EncodingConfig config;
-    // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-    config.curveFrequencyLimit = 72.5;
-    // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-    config.frequencyBandLimits = std::vector<std::pair<double, double>>{{72.5, 1000}};
-    // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-    config.windowLength = 15;
-    // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-    config.wavelet_windowLength = 128;
-    // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
-    config.wavelet_bitbudget = 4;
+    if (inputParser.cmdOptionExists("-kb")) {
+      int bitrate = std::stoi(inputParser.getCmdOption("-kb"));
+      std::cout << "target bitrate: " << bitrate << " kb/s" << std::endl;
+      config = haptics::encoder::EncodingConfig::generateConfig(bitrate);
+    } else {
+      config = haptics::encoder::EncodingConfig::generateConfig();
+    }
     codeExit = PcmEncoder::encode(filename, config, myPerception);
     hapticFile.addPerception(myPerception);
-  }
-  else {
+  } else {
     codeExit = EXIT_FAILURE;
   }
 
   if (codeExit == EXIT_FAILURE) {
-    InputParser::help(args[0]);
+    help();
     return codeExit;
   }
 
-  IOJson::writeFile(hapticFile, output);
+  if (inputParser.cmdOptionExists("-b") || inputParser.cmdOptionExists("--binary")) {
+    IOBinary::writeFile(hapticFile, output);
+  } else {
+    IOJson::writeFile(hapticFile, output);
+  }
+
   return codeExit;
 }
